enabled running test groups from testng.xml
