+++ src/test/java/testClasses/SearchTest.java
@@ -5,10 +5,7 @@ import com.aventstack.extentreports.Status;
 import org.testng.annotations.Parameters;
 import org.testng.annotations.Test;
 import pageObjects.SearchPage;
-import testUtilities.extentReportsClasses.ExtentReporterManager;
-import testUtilities.extentReportsClasses.ExtentTestManager;
 
-import java.lang.reflect.Method;
 
 import static org.testng.Assert.assertFalse;
 import static org.testng.Assert.assertTrue;
@@ -20,8 +17,8 @@ public class SearchTest extends BaseTest {
     public void testSearchForItem(String itemForSearch) {
 
         SearchPage sp = hp.searchForAnItem(itemForSearch);
-        String searchPageTitle = sp.getPageTitle();
-        assertFalse(searchPageTitle.contains(itemForSearch));
+        String searchPageTitle = sp.getSearchPageTitle();
+        assertTrue(searchPageTitle.contains(itemForSearch));
       }
 
 }
